Confirm that the issue is only during unload.
I have confirmed that the zmq server does work and receive mesages.
This can be done by uncommenting the block of code before the semaphore
wait.
It will send a hello message every second.
The issue seems to be only that I cannot close the socket in any way
short of shutting down VBS3.